fix display
